Time: 7 ms (43.12%), Space: 15.7 MB (75%) - LeetHub

+++ 3493-maximum-number-of-operations-to-move-ones-to-the-end/3493-maximum-number-of-operations-to-move-ones-to-the-end.cpp
@@ -10,7 +10,7 @@ class Solution {
       ans += ones;
 
  
-    return ans;
+    return (int)ans;
   }
 };
